requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.5
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.34.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
twilio==9.1.1
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.4.post2
